CI/CD github actions now working
Was wondering why github actions was not quite working so made a hard fork, but that still didn't fix it. Turns out this old repo was using master not main, so hence it didn't trigger on commit.

This commit fixes this oversight. But anyway the hard fork is okay, as the original maintainer appears to have stopped maintaining it anyhow.