v1.1.2 Fix edge case where there is more KV pair in query string than max param
@nike80 found out a bug and commented in https://github.com/jacketizer/libyuarel/issues/16

> Hello, I found an error when 1 struct yuarel_param is passed to the yuarel_parse_query() function. They are not working properly.
>
> ```
> 		char* query = "a=b&c=d";
> 		yuarel_param params;
> 		yuarel_parse_query(query, '&', &params, 1);
>
> 		// HERE
> 		// params.key == "a"
> 		// params.val ==  "b&c=d"
> ```

My investigation traced a bug that I found and can explain as such below:

This is actually because in https://github.com/jacketizer/libyuarel/blob/master/yuarel.c#L274 it only splits the `&` delimiter up to max_param times. To fix this issue we would need to do a clean up step where we run this one more time to check if there is additional `&` to at least null terminate.

This PR will fix this particular issue and adds a few new test cases to catch it later.